// src/types/ViewType.ts
export enum ViewType {
  MAIN = "main",
  ORDERS = "orders",
  DATA = "data",
  BENEFITS = "benefits",
  PAYMENT_METHOD = "payment-method",
  CANCEL_REASON = "cancel-reason",
  CANCEL_OFFER = "cancel-offer",
  CANCEL_CONFIRM = "cancel-confirm",
  CANCEL_SUCCESS = "cancel-success",
  BENEFIT_SUCCESS = "benefit-success",
}